refactor: refactored release page
